Use React-Router and Bulma
This is too large a commit... I'll try and do better with smaller,
focused commits in the future.

Add both `Bulma` and `react-router`. Start creating, and testing,
routing. We introduce the main components.

We get rid of any notion of a "complex" form for now, preferring to
instead just render simple components. We will build out richer
components over time.