Split up the js into multiple files
Previously, we only had one huge `App.js` file, which would've quickly
gotten out of hand.

We address this by creating a `components` directory and a `services`
directory. The `components` directory is for React components - it
should ONLY relate to UI and there should be NO complex logic or
interaction with the backend (outside of calling service objects which will
communicate with the backend). We also create a `services` directory
which is for defining service objects for any non-trivial logic/backend
interactions. At the very least, we will have a `backend` service for
backend interactions. We will pass services to react components (which
should also make testing individual components much simpler).