Add prettier and eslint
Add prettier and eslint for front-end. Both can be run via `yarn`
scripts, or via `yarn run ...`. We install `prettier` via `yarn add`. Do
not install `eslint` manually - the `create-react-app` template already
has it, and is opinionated about the version.

We still need to setup `vscode` to use `prettier` on save.

We will also want to make running the linter and a formatter check as
part of travis. We also need to fix the existing linter errors.