Fix broken routing
I believe that the switch from `BrowserRouter` to `Router
history=createBrowserHistory()` broke our ability to actually navigate
via the web.

However, the change back to `BrowserHistory` would break our test suite.

Address by passing in the `browserComponent` as an argument to the `App`
argument. This lets me use one approach in testing and a different one
when running the actual app.

I punted a little bit on the types here... it was getting a little
confusing w.r.t. the proper type for the injected browserComponent. I
may revisit, or there may be a more elegant way to do this entirely.