[WIP] MMVP front end
Getting the front end to a place where it can at least be hosted on the
internet.

Using https://bulmatemplates.github.io/bulma-templates/templates/cover.html
as a sample template.

Start passing in the `navbar` and `footer` components to the `page`
components, so the `page` components can be in control of how they are
rendered.

Still todo:

- Determine if we should update our naming to indicate the `home`,
`about`, and `makeAMentalHealthPlan` are top level "page" components.
- Update the other pages (/about, /make-a-mental-health-plan) w/ the proper content. Can be _super_
lightweight.
- Replace the bulma logo in the navbar.
- Replace the random image w/ an actual image.